# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AegisOps AI
  description: AI-powered autonomous SRE platform with anomaly detection, incident
    investigation, and human-approved remediation.
  version: 1.0.0
paths:
...
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get Me
      operationId: get_me_api_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
...
  /api/incidents/{incident_id}:
    get:
      tags:
      - Incident Management
      summary: Get Incident
      operationId: get_incident_api_incidents__incident_id__get
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Incident Api Incidents  Incident Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/incidents/{incident_id}/investigate:
    post:
      tags:
      - Incident Management
      summary: Investigate Incident
      description: Executes the full multi-agent investigation workflow.
      operationId: investigate_incident_api_incidents__incident_id__investigate_post
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/incidents/{incident_id}/remediation/approve:
    post:
      tags:
      - Incident Management
      summary: Approve Remediation
      description: Executes human-approved remediation in simulator, verifies recovery,
        and drafts postmortem.
      operationId: approve_remediation_api_incidents__incident_id__remediation_approve_post
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemediationApprovalRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/incidents/{incident_id}/postmortem:
    get:
      tags:
      - Incident Management
      summary: Get Postmortem
      operationId: get_postmortem_api_incidents__incident_id__postmortem_get
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    RemediationApprovalRequest:
      properties:
        action_id:
          type: string
          title: Action Id
        approved:
          type: boolean
          title: Approved
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - action_id
      - approved
      title: RemediationApprovalRequest
...
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          $ref: '#/components/schemas/UserResponse'
      type: object
      required:
      - access_token
      - user
      title: Token
...
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          $ref: '#/components/schemas/UserRole'
          default: OPERATOR
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - full_name
      - password
      title: UserCreate
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          $ref: '#/components/schemas/UserRole'
          default: OPERATOR
        id:
          type: integer
          title: Id
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - email
      - full_name
      - id
      - is_active
      - created_at
      title: UserResponse
    UserRole:
      type: string
      enum:
      - ADMIN
      - OPERATOR
      - VIEWER
      title: UserRole
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer